refactor(adaptive result): replace footer to components
