+++ src/components/Footer/Footer.module.scss
@@ -1,4 +1,4 @@
-@import "../../../app/styles/mixins";
+@import "../../app/styles/mixins";
 
 .footer {
   @include section(row);
