+++ src/pages/FavouritePage/ui/FavouritePage.tsx
@@ -3,7 +3,7 @@ import styles from "./FavouritePage.module.scss";
 import Header from "../../../components/ui/Header/Header";
 import FavouriteNavigate from "./FavouriteNavigate/FavouriteNavigate";
 import FavouriteList from "./FavouriteList/FavouriteList";
-import Footer from "../../../components/ui/Footer/Footer";
+import Footer from "../../../components/Footer/Footer";
 
 const FavouritePage: FC = () => {
   return (
