+++ src/pages/Login/Login.tsx
@@ -8,7 +8,7 @@ import { IconColor, IconTypes } from "../../components/ui/Icon/types";
 import SearchHeader from "../SearchPage/ui/SearchHeader/SearchHeader";
 import Input from "../../components/ui/Input/Input";
 import AuthBannerForm from "../../components/ui/AuthBannerForm/AuthBannerForm";
-import Footer from "../../components/ui/Footer/Footer";
+import Footer from "../../components/Footer/Footer";
 
 interface ILogin {
   email: string;
